Separate train_model into functions and main
